Feat

+++ src/main/java/com/compass/ecommerce/Service/ProductService.java
@@ -1,18 +1,33 @@
 package com.compass.ecommerce.Service;
 
-import com.compass.ecommerce.DTO.ProductDTORequest;
-import com.compass.ecommerce.DTO.ProductDTOResponse;
+import com.compass.ecommerce.DTO.Product.GetProductDTOResponse;
+import com.compass.ecommerce.DTO.Product.ProductDTORequest;
+import com.compass.ecommerce.DTO.Product.ProductDTOResponse;
+import com.compass.ecommerce.DTO.Product.UpdateStockDTORequest;
 import com.compass.ecommerce.Exception.ResourceNotFoundException;
 import com.compass.ecommerce.model.Product;
+import com.compass.ecommerce.model.Sale;
+import com.compass.ecommerce.model.enums.SaleStatus;
+import com.compass.ecommerce.repository.ItemSaleRepository;
 import com.compass.ecommerce.repository.ProductRepository;
+import com.compass.ecommerce.repository.SaleRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @Transactional
 public class ProductService {
 
     private ProductRepository productRepository;
+
+    @Autowired
+    private SaleRepository saleRepository;
+
+    @Autowired
+    private ItemSaleRepository itemSaleRepository;
 
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
@@ -24,6 +39,7 @@ public class ProductService {
         product.setDescription(productDTORequest.description());
         product.setPrice(productDTORequest.price());
         product.setQuantity(productDTORequest.quantity());
+        product.setStock(productDTORequest.stock());
 
         productRepository.save(product);
 
@@ -32,14 +48,15 @@ public class ProductService {
                 product.getName(),
                 product.getDescription(),
                 product.getPrice(),
-                product.getQuantity());
+                product.getQuantity(),
+                product.getStock());
     }
 
-    public ProductDTOResponse getProductById(Long id) {
+    public GetProductDTOResponse getProductById(Long id) {
         Product product = productRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " +id));
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + id));
 
-        return new ProductDTOResponse(
+        return new GetProductDTOResponse(
                 product.getId(),
                 product.getName(),
                 product.getDescription(),
@@ -49,12 +66,13 @@ public class ProductService {
 
     public ProductDTOResponse updateProduct(Long id, ProductDTORequest productDTORequest) {
         Product product = productRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " +id));
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + id));
 
         product.setName(productDTORequest.name());
         product.setDescription(productDTORequest.description());
         product.setPrice(productDTORequest.price());
         product.setQuantity(productDTORequest.quantity());
+        product.setStock(productDTORequest.quantity());
 
         productRepository.save(product);
 
@@ -62,12 +80,60 @@ public class ProductService {
                 product.getName(),
                 product.getDescription(),
                 product.getPrice(),
-                product.getQuantity());
+                product.getQuantity(),
+                product.getStock()
+        );
+    }
+
+    // Adiciona produtos ao estoque
+    public ProductDTOResponse addProductStock(Long id ,UpdateStockDTORequest addProduct) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + id));
+
+        product.addStock(addProduct.quantity());
+
+        productRepository.save(product);
+
+        return new ProductDTOResponse(
+                product.getId(),
+                product.getName(),
+                product.getDescription(),
+                product.getPrice(),
+                product.getQuantity(),
+                product.getStock()
+        );
+    }
+
+    // Remove produtos do estoque
+    public ProductDTOResponse removeProductStock(Long id, UpdateStockDTORequest removeStock) {
+        Product product = productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + id));
+
+        product.removeFromStock(removeStock.quantity());
+
+        productRepository.save(product);
+
+        return new ProductDTOResponse(
+                product.getId(),
+                product.getName(),
+                product.getDescription(),
+                product.getPrice(),
+                product.getQuantity(),
+                product.getStock()
+        );
     }
 
     public void deleteProduct(Long id) {
         Product product = productRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " +id));
+                .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + id));
+
+
+        List<Sale> salesWithProduct = saleRepository.findAllByProductIdAndStatus(id, SaleStatus.PROCESSING);
+        if (!salesWithProduct.isEmpty()) {
+            throw new ResourceNotFoundException("Não é possível excluir o produto enquanto houver vendas em processamento associadas.");
+        }
+
+        itemSaleRepository.deleteByProductId(id);
 
         productRepository.delete(product);
     }
