+++ src/main/java/com/compass/ecommerce/Service/SaleService.java
@@ -1,23 +1,29 @@
 package com.compass.ecommerce.Service;
 
-import com.compass.ecommerce.DTO.ItemSaleDTOResponse;
-import com.compass.ecommerce.DTO.SaleDTORequest;
-import com.compass.ecommerce.DTO.SaleDTOResponse;
+import com.compass.ecommerce.DTO.ItemSale.ItemSaleDTOResponse;
+import com.compass.ecommerce.DTO.Sale.SaleDTORequest;
+import com.compass.ecommerce.DTO.Sale.SaleDTOResponse;
 import com.compass.ecommerce.Exception.ResourceNotFoundException;
 import com.compass.ecommerce.model.ItemSale;
 import com.compass.ecommerce.model.Product;
 import com.compass.ecommerce.model.Sale;
 import com.compass.ecommerce.model.User;
+import com.compass.ecommerce.model.enums.SaleStatus;
 import com.compass.ecommerce.repository.ItemSaleRepository;
 import com.compass.ecommerce.repository.ProductRepository;
 import com.compass.ecommerce.repository.SaleRepository;
 import com.compass.ecommerce.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.time.DayOfWeek;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -34,37 +40,36 @@ public class SaleService {
 
     public SaleDTOResponse createSale(SaleDTORequest saleDTORequest) {
         User user = userRepository.findById(saleDTORequest.userId())
-                .orElseThrow(() -> new ResourceNotFoundException("Código do cliente inválido"));
+                .orElseThrow(() -> new ResourceNotFoundException("Código do usuário inválido"));
 
 
         Sale sale = new Sale();
         sale.setUser(user);
         sale.setDate(LocalDateTime.now());
+        sale.setStatus(SaleStatus.PROCESSING);
 
         //Converte os itens da requisição em uma lista de ItemSale
         List<ItemSale> itemSales = saleDTORequest.items().stream()
                 .map(itemDTO -> {
                     // Verifica se o ID do produto é nulo
                     if (itemDTO.productId() == null) {
-                        throw new IllegalArgumentException("O id do produto não pode ser nulo");
+                        throw new ResourceNotFoundException("O id do produto não pode ser nulo");
                     }
 
                     // Busca o produto pelo ID
                     Product product = productRepository.findById(itemDTO.productId())
                             .orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado pelo id: " + itemDTO.productId()));
 
+                    System.out.println(product.getStock());
 
-                    if (product.getQuantity() < itemDTO.quantity()){
-                        throw new RuntimeException("Estoque insuficiente do produto: " +product.getName());
-                    }
-
+                    product.removeFromStock(itemDTO.stock());
 
                     // Cria um novo ItemSale e o associa ao produto e à venda
                     ItemSale itemSale = new ItemSale();
 
                     itemSale.setProduct(product);
 
-                    itemSale.setQuantity(itemDTO.quantity());
+                    itemSale.setQuantity(itemDTO.stock());
                     itemSale.setSale(sale);
 
                     return itemSale;
@@ -72,7 +77,8 @@ public class SaleService {
                 })
                 .collect(Collectors.toList());
 
-        for(ItemSale itemSale :itemSales){
+        // Lógica para controle do estoque
+        for (ItemSale itemSale : itemSales) {
             Product product = itemSale.getProduct();
 
             Integer newQuantity = product.getQuantity() - itemSale.getQuantity();
@@ -97,6 +103,98 @@ public class SaleService {
                 sale.getId(),
                 sale.getUser().getName(),
                 sale.getDate(),
+                sale.getStatus(),
+                sale.getTotal(),
+                sale.getItemSales().stream()
+                        .map(itemSale -> new ItemSaleDTOResponse(
+                                itemSale.getProduct().getId(),
+                                itemSale.getProduct().getName(),
+                                itemSale.getProduct().getDescription(),
+                                itemSale.getQuantity()
+                        ))
+                        .collect(Collectors.toList())
+        );
+    }
+
+    public SaleDTOResponse confirmSale(Long id) {
+        Sale sale = saleRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Venda não encontrada pelo id: " + id));
+
+        // Pega o usuário autenticado
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        User authenticatedUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!sale.getUser().getId().equals(authenticatedUser.getId()) && !isAdmin) {
+            throw new ResourceNotFoundException("Você não tem permissão para confirmar esta venda");
+        }
+
+        if (sale.getStatus() == SaleStatus.CANCELED) {
+            throw new ResourceNotFoundException("Não é possível finalizar uma venda que já foi cancelada.");
+        }
+
+        sale.setStatus(SaleStatus.FINISHED);
+        saleRepository.save(sale);
+
+        return new SaleDTOResponse(
+                sale.getId(),
+                sale.getUser().getName(),
+                sale.getDate(),
+                sale.getStatus(),
+                sale.getTotal(),
+                sale.getItemSales().stream()
+                        .map(itemSale -> new ItemSaleDTOResponse(
+                                itemSale.getProduct().getId(),
+                                itemSale.getProduct().getName(),
+                                itemSale.getProduct().getDescription(),
+                                itemSale.getQuantity()
+                        ))
+                        .collect(Collectors.toList())
+        );
+    }
+
+    public SaleDTOResponse cancelSale(Long id) {
+        Sale sale = saleRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Venda não encontrada pelo id: " + id));
+
+        // Pega o usuário autenticado
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        User authenticatedUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!sale.getUser().getId().equals(authenticatedUser.getId()) && !isAdmin) {
+            throw new ResourceNotFoundException("Você não tem permissão para cancelar esta venda");
+        }
+
+        if (sale.getStatus() == SaleStatus.FINISHED) {
+            throw new ResourceNotFoundException("Não é possível cancelar uma venda que já foi finalizada.");
+        }
+
+        sale.setStatus(SaleStatus.CANCELED);
+
+        for (ItemSale itemSale : sale.getItemSales()) {
+            Product product = itemSale.getProduct();
+            Integer restoredStock = product.getQuantity() + itemSale.getQuantity();
+            product.setQuantity(restoredStock);
+            product.setStock(product.updateStock(itemSale.getQuantity()));
+            productRepository.save(product);
+        }
+
+        saleRepository.save(sale);
+
+        return new SaleDTOResponse(
+                sale.getId(),
+                sale.getUser().getName(),
+                sale.getDate(),
+                sale.getStatus(),
                 sale.getTotal(),
                 sale.getItemSales().stream()
                         .map(itemSale -> new ItemSaleDTOResponse(
@@ -111,7 +209,7 @@ public class SaleService {
 
     public SaleDTOResponse findById(Long id) {
         Sale sale = saleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Venda não encontrada pelo id: " +id));
+                .orElseThrow(() -> new ResourceNotFoundException("Venda não encontrada pelo id: " + id));
 
         List<ItemSaleDTOResponse> items = sale.getItemSales().stream()
                 .map(itemSale -> new ItemSaleDTOResponse(
@@ -127,6 +225,7 @@ public class SaleService {
                 sale.getId(),
                 sale.getUser().getName(),
                 sale.getDate(),
+                sale.getStatus(),
                 sale.getTotal(),
                 items);
 
@@ -134,22 +233,38 @@ public class SaleService {
 
     public SaleDTOResponse updateSale(Long id, SaleDTORequest saleDTORequest) {
         Sale sale = saleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Venda não encontrada pelo id: " +id));
+                .orElseThrow(() -> new ResourceNotFoundException("Venda não encontrada pelo id: " + id));
 
-        User user = userRepository
-                .findById(saleDTORequest.userId())
-                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado pelo ID: " + saleDTORequest.userId()));
+        // Pega o usuário autenticado
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        User authenticatedUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
 
-        sale.setUser(user);
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!sale.getUser().getId().equals(authenticatedUser.getId()) && !isAdmin) {
+            throw new ResourceNotFoundException("Você não tem permissão para Atualizar esta venda");
+        }
+
+        if (sale.getStatus() == SaleStatus.FINISHED) {
+            throw new ResourceNotFoundException("Não é possível atualizar uma venda que já foi finalizada.");
+        }
+
+        if (sale.getStatus() == SaleStatus.CANCELED) {
+            throw new ResourceNotFoundException("Não é possível atualizar uma venda que já foi cancelada.");
+        }
 
         sale.setDate(LocalDateTime.now());
+
         //Remove todos os itens associados à venda
         sale.getItemSales().forEach(itemSale -> itemSaleRepository.delete(itemSale));
 
         List<ItemSale> updatedItemSales = saleDTORequest.items().stream()
                 .map(itemDTO -> {
                     if (itemDTO.productId() == null) {
-                        throw new IllegalArgumentException("O id do produto não pode ser nulo");
+                        throw new ResourceNotFoundException("O id do produto não pode ser nulo");
                     }
 
                     Product product = productRepository.findById(itemDTO.productId())
@@ -157,12 +272,14 @@ public class SaleService {
 
                     ItemSale itemSale = new ItemSale();
                     itemSale.setProduct(product);
-                    itemSale.setQuantity(itemDTO.quantity());
+                    itemSale.setQuantity(itemDTO.stock());
                     itemSale.setSale(sale);
 
                     return itemSale;
                 })
                 .collect(Collectors.toList());
+
+        sale.setStatus(sale.getStatus());
 
         itemSaleRepository.saveAll(updatedItemSales);
         sale.setItemSales(updatedItemSales);
@@ -179,6 +296,7 @@ public class SaleService {
                 sale.getId(),
                 sale.getUser().getName(),
                 sale.getDate(),
+                sale.getStatus(),
                 sale.getTotal(),
                 sale.getItemSales().stream()
                         .map(itemSale -> new ItemSaleDTOResponse(
@@ -193,11 +311,40 @@ public class SaleService {
 
     public void deleteSale(Long id) {
         Sale sale = saleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Venda não encontrada pelo id: " +id));
+                .orElseThrow(() -> new RuntimeException("Venda não encontrada pelo id: " + id));
 
         //Remove todos os itens associados à venda
         sale.getItemSales().forEach(itemSale -> itemSaleRepository.delete(itemSale));
 
         saleRepository.delete(sale);
     }
+
+    public List<SaleDTOResponse> weeklyReport(LocalDate date) {
+        // Calcular o início e o fim da semana
+        LocalDateTime startWeek = date.with(DayOfWeek.MONDAY).atStartOfDay();  // Início da semana (segunda-feira)
+        LocalDateTime endWeek = date.with(DayOfWeek.SUNDAY).atTime(LocalTime.MAX);  // Fim da semana (domingo)
+
+        // Buscar as vendas no intervalo entre o início e o fim da semana
+        List<Sale> sales = saleRepository.findSalesByDateRange(startWeek, endWeek);
+
+        // Converter as vendas para DTOs de resposta
+        return sales.stream()
+                .map(sale -> new SaleDTOResponse(
+                        sale.getId(),
+                        sale.getUser().getName(),
+                        sale.getDate(),
+                        sale.getStatus(),
+                        sale.getTotal(),
+                        sale.getItemSales().stream()
+                                .map(itemSale -> new ItemSaleDTOResponse(
+                                        itemSale.getProduct().getId(),
+                                        itemSale.getProduct().getName(),
+                                        itemSale.getProduct().getDescription(),
+                                        itemSale.getQuantity()
+                                ))
+                                .collect(Collectors.toList())
+                ))
+                .collect(Collectors.toList());
+    }
+
 }
